requirements.txt
numpy==1.26.4
pillow==11.1.0
gradio==5.23.1
ultralytics==8.3.98
opencv-python==4.11.0.86
# torch>=2.2.2

